🚀feat(v9-actionFn) add project to local state
This commit adds a new project object to the local state. The project object is created with the command's args, an id, and a temp_id. The default project object is defined, and then it is merged with the project object to add. The added project is then pushed to the "projects" array in the local state. Finally, the updated local state is returned.